+++ transports/dns/src/lib.rs
@@ -75,19 +75,15 @@ pub mod async_std {
         /// Creates a new [`Transport`] from the OS's DNS configuration and defaults.
         pub async fn system(inner: T) -> Result<Transport<T>, io::Error> {
             let (cfg, opts) = system_conf::read_system_conf()?;
-            Self::custom(inner, cfg, opts).await
+            Ok(Self::custom(inner, cfg, opts).await)
         }
 
         /// Creates a [`Transport`] with a custom resolver configuration and options.
-        pub async fn custom(
-            inner: T,
-            cfg: ResolverConfig,
-            opts: ResolverOpts,
-        ) -> Result<Transport<T>, io::Error> {
-            Ok(Transport {
+        pub async fn custom(inner: T, cfg: ResolverConfig, opts: ResolverOpts) -> Transport<T> {
+            Transport {
                 inner: Arc::new(Mutex::new(inner)),
                 resolver: async_std_resolver::resolver(cfg, opts).await,
-            })
+            }
         }
     }
 }
@@ -108,9 +104,9 @@ pub mod tokio {
 
     impl<T> Transport<T> {
         /// Creates a new [`Transport`] from the OS's DNS configuration and defaults.
-        pub fn system(inner: T) -> Result<crate::Transport<T, TokioAsyncResolver>, std::io::Error> {
+        pub fn system(inner: T) -> Result<Transport<T>, std::io::Error> {
             let (cfg, opts) = system_conf::read_system_conf()?;
-            Self::custom(inner, cfg, opts)
+            Ok(Self::custom(inner, cfg, opts))
         }
 
         /// Creates a [`Transport`] with a custom resolver configuration
@@ -119,12 +115,11 @@ pub mod tokio {
             inner: T,
             cfg: trust_dns_resolver::config::ResolverConfig,
             opts: trust_dns_resolver::config::ResolverOpts,
-        ) -> Result<crate::Transport<T, TokioAsyncResolver>, std::io::Error> {
-            // TODO: Make infallible in next breaking release. Or deprecation?
-            Ok(Transport {
+        ) -> Transport<T> {
+            Transport {
                 inner: Arc::new(Mutex::new(inner)),
                 resolver: TokioAsyncResolver::tokio(cfg, opts),
-            })
+            }
         }
     }
 }
@@ -780,9 +775,7 @@ mod tests {
                 .build()
                 .unwrap();
 
-            rt.block_on(run(
-                tokio::Transport::custom(CustomTransport, config, opts).unwrap()
-            ));
+            rt.block_on(run(tokio::Transport::custom(CustomTransport, config, opts)));
         }
     }
 }